fix(test_utils): do not ignore unused imports, closes #1102
