+++ aries/misc/legacy/libvdrtools/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub use indy_utils::environment;
 pub mod crypto;
 pub use indy_utils::wql;
 #[macro_use]
