+++ aries/misc/test_utils/src/devsetup.rs
@@ -1,22 +1,13 @@
-#![allow(unused_imports)]
-
 use std::{
-    env,
-    error::Error,
     fs::{self, DirBuilder, OpenOptions},
-    future::Future,
     io::Write,
     path::{Path, PathBuf},
-    sync::{Arc, Once},
 };
 
 use agency_client::testing::mocking::{enable_agency_mocks, AgencyMockDecrypted};
-#[cfg(feature = "credx")]
-use aries_vcx_core::anoncreds::credx_anoncreds::IndyCredxAnonCreds;
 use aries_vcx_core::{
     anoncreds::base_anoncreds::BaseAnonCreds,
     errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult},
-    global::settings::{DEFAULT_WALLET_KEY, WALLET_KDF_RAW},
     ledger::{
         base_ledger::{
             AnoncredsLedgerRead, AnoncredsLedgerWrite, IndyLedgerRead, IndyLedgerWrite,
@@ -24,36 +15,27 @@ use aries_vcx_core::{
         },
         indy::pool::test_utils::{create_testpool_genesis_txn_file, get_temp_file_path},
         indy_vdr_ledger::{
-            build_ledger_components, indyvdr_build_ledger_read, indyvdr_build_ledger_write,
-            DefaultIndyLedgerRead, DefaultIndyLedgerWrite, GetTxnAuthorAgreementData,
-            IndyVdrLedgerRead, IndyVdrLedgerReadConfig, IndyVdrLedgerWrite,
-            IndyVdrLedgerWriteConfig, ProtocolVersion, VcxPoolConfig,
+            build_ledger_components, DefaultIndyLedgerRead, DefaultIndyLedgerWrite,
+            GetTxnAuthorAgreementData, VcxPoolConfig,
         },
-        request_submitter::vdr_ledger::{IndyVdrLedgerPool, IndyVdrSubmitter},
-        response_cacher::in_memory::{InMemoryResponseCacher, InMemoryResponseCacherConfig},
     },
-    wallet::{base_wallet::BaseWallet, mock_wallet::MockWallet},
-    PoolConfig, ResponseParser,
-};
-#[cfg(feature = "vdr_proxy_ledger")]
-use aries_vcx_core::{ledger::request_submitter::vdr_proxy::VdrProxySubmitter, VdrProxyClient};
-#[cfg(feature = "vdrtools_wallet")]
-use aries_vcx_core::{
-    wallet::indy::{
-        wallet::{create_and_open_wallet, create_and_store_my_did},
-        WalletConfig,
-    },
-    WalletHandle,
+    wallet::base_wallet::BaseWallet,
 };
 use chrono::{DateTime, Duration, Utc};
-use lazy_static::lazy_static;
 use libvcx_logger::init_test_logging;
 use log::{debug, info, warn};
 
-use crate::{
-    constants::{INSTITUTION_DID, POOL1_TXN, TRUSTEE_SEED},
-    mockdata::{mock_anoncreds::MockAnoncreds, mock_ledger::MockLedger},
-};
+use crate::constants::{POOL1_TXN, TRUSTEE_SEED};
+
+#[cfg(feature = "vdr_proxy_ledger")]
+pub mod vdr_proxy_ledger;
+#[cfg(feature = "vdr_proxy_ledger")]
+use crate::devsetup::vdr_proxy_ledger::dev_build_profile_vdr_proxy_ledger;
+
+#[cfg(feature = "vdrtools_wallet")]
+pub mod vdrtools_wallet;
+#[cfg(feature = "vdrtools_wallet")]
+use crate::devsetup::vdrtools_wallet::dev_build_indy_wallet;
 
 const DEFAULT_AML_LABEL: &str = "eula";
 
@@ -167,27 +149,6 @@ impl Drop for SetupMocks {
     }
 }
 
-#[cfg(feature = "vdrtools_wallet")]
-pub async fn dev_setup_wallet_indy(key_seed: &str) -> (String, WalletHandle) {
-    info!("dev_setup_wallet_indy >>");
-    let config_wallet = WalletConfig {
-        wallet_name: format!("wallet_{}", uuid::Uuid::new_v4()),
-        wallet_key: DEFAULT_WALLET_KEY.into(),
-        wallet_key_derivation: WALLET_KDF_RAW.into(),
-        wallet_type: None,
-        storage_config: None,
-        storage_credentials: None,
-        rekey: None,
-        rekey_derivation_method: None,
-    };
-    let wallet_handle = create_and_open_wallet(&config_wallet).await.unwrap();
-    // todo: can we just extract this away? not always we end up using it (alice test agent)
-    let (did, _vk) = create_and_store_my_did(wallet_handle, Some(key_seed), None)
-        .await
-        .unwrap();
-
-    (did, wallet_handle)
-}
 pub fn dev_build_profile_vdr_ledger(
     genesis_file_path: String,
 ) -> (DefaultIndyLedgerRead, DefaultIndyLedgerWrite) {
@@ -203,45 +164,6 @@ pub fn dev_build_profile_vdr_ledger(
     (ledger_read, ledger_write)
 }
 
-#[cfg(feature = "vdr_proxy_ledger")]
-pub async fn dev_build_profile_vdr_proxy_ledger() -> (
-    IndyVdrLedgerRead<VdrProxySubmitter, InMemoryResponseCacher>,
-    IndyVdrLedgerWrite<VdrProxySubmitter>,
-) {
-    info!("dev_build_profile_vdr_proxy_ledger >>");
-
-    let client_url =
-        env::var("VDR_PROXY_CLIENT_URL").unwrap_or_else(|_| "http://127.0.0.1:3030".to_string());
-    let client = VdrProxyClient::new(&client_url).unwrap();
-
-    let request_submitter = VdrProxySubmitter::new(Arc::new(client));
-    let response_parser = ResponseParser;
-    let cacher_config = InMemoryResponseCacherConfig::builder()
-        .ttl(std::time::Duration::from_secs(60))
-        .capacity(1000)
-        .unwrap()
-        .build();
-    let response_cacher = InMemoryResponseCacher::new(cacher_config);
-
-    let config_read = IndyVdrLedgerReadConfig {
-        request_submitter: request_submitter.clone(),
-        response_parser,
-        response_cacher,
-        protocol_version: ProtocolVersion::Node1_4,
-    };
-    let ledger_read = IndyVdrLedgerRead::new(config_read);
-
-    let config_write = IndyVdrLedgerWriteConfig {
-        request_submitter,
-        taa_options: prepare_taa_options(&ledger_read).await.unwrap(),
-        protocol_version: ProtocolVersion::Node1_4,
-    };
-    let ledger_write = IndyVdrLedgerWrite::new(config_write);
-
-    (ledger_read, ledger_write)
-}
-
-#[allow(unreachable_code)]
 #[allow(unused_variables)]
 pub async fn dev_build_featured_indy_ledger(
     genesis_file_path: String,
@@ -262,25 +184,21 @@ pub async fn dev_build_featured_indy_ledger(
     };
 }
 
-#[cfg(feature = "vdrtools_wallet")]
-pub async fn dev_build_indy_wallet(key_seed: &str) -> (String, impl BaseWallet) {
-    use aries_vcx_core::wallet::indy::IndySdkWallet;
-
-    let (public_did, wallet_handle) = dev_setup_wallet_indy(key_seed).await;
-    (public_did, IndySdkWallet::new(wallet_handle))
-}
-
-#[allow(unreachable_code)]
-#[allow(unused_variables)]
+#[allow(clippy::needless_return)]
 pub async fn dev_build_featured_anoncreds() -> impl BaseAnonCreds {
     #[cfg(feature = "credx")]
-    return IndyCredxAnonCreds;
+    {
+        use aries_vcx_core::anoncreds::credx_anoncreds::IndyCredxAnonCreds;
+        return IndyCredxAnonCreds;
+    }
 
     #[cfg(not(feature = "credx"))]
-    return MockAnoncreds;
+    return {
+        use crate::mockdata::mock_anoncreds::MockAnoncreds;
+        return MockAnoncreds;
+    };
 }
 
-#[allow(unreachable_code)]
 #[allow(unused_variables)]
 pub async fn dev_build_featured_wallet(key_seed: &str) -> (String, impl BaseWallet) {
     #[cfg(feature = "vdrtools_wallet")]
@@ -290,7 +208,13 @@ pub async fn dev_build_featured_wallet(key_seed: &str) -> (String, impl BaseWall
     };
 
     #[cfg(not(feature = "vdrtools_wallet"))]
-    return (INSTITUTION_DID.to_owned(), MockWallet);
+    {
+        use aries_vcx_core::wallet::mock_wallet::MockWallet;
+
+        use crate::constants::INSTITUTION_DID;
+
+        return (INSTITUTION_DID.to_owned(), MockWallet);
+    }
 }
 
 pub async fn build_setup_profile() -> SetupProfile<
@@ -299,8 +223,6 @@ pub async fn build_setup_profile() -> SetupProfile<
     impl BaseAnonCreds,
     impl BaseWallet,
 > {
-    use aries_vcx_core::anoncreds::base_anoncreds::BaseAnonCreds;
-
     init_test_logging();
 
     let genesis_file_path = get_temp_file_path(POOL1_TXN).to_str().unwrap().to_string();
